+++ packages/wp-mu-plugin/ionos-core/ionos-core.php
@@ -15,3 +15,79 @@ namespace ionos\ionos_core;
 defined('ABSPATH') || exit();
 
 require_once __DIR__ . '/ionos-core/index.php';
+
+
+const INFO_JSON_URL   = 'https://tom-rockstar.de/ionos-core/ionos-core-info.json';
+const CURRENT_VERSION = '0.1.0';
+
+\add_action('wp_update_plugins', function (): void
+{
+  $info = \wp_remote_get(INFO_JSON_URL, ['timeout' => 5]);
+
+  if (\is_wp_error($info)) {
+    \error_log('ionos-core: Error fetching update info: ' . $info->get_error_message());
+    return;
+  }
+
+  $info_data    = json_decode(\wp_remote_retrieve_body($info), true, 512, JSON_THROW_ON_ERROR);
+  $latest       = $info_data['version'] ?? null;
+  $download_url = $info_data['download_url'] ?? null;
+
+  if (!$latest || !$download_url) {
+    \error_log('ionos-core: Update info response is missing version or download_url.');
+    return;
+  }
+
+  if (!\version_compare($latest, CURRENT_VERSION, '>')) {
+    return;
+  }
+
+  \error_log('ionos-core: Updating from ' . CURRENT_VERSION . ' to ' . $latest . '.');
+
+  $result = (new MU_Plugin_Upgrader())->upgrade($download_url);
+
+  if (\is_wp_error($result)) {
+    \error_log('ionos-core: Update failed: ' . $result->get_error_message());
+    return;
+  }
+
+  \error_log('ionos-core: Update to ' . $latest . ' completed successfully.');
+});
+
+
+require_once ABSPATH . 'wp-admin/includes/class-wp-upgrader.php';
+require_once ABSPATH . 'wp-admin/includes/file.php';
+
+class MU_Plugin_Upgrader extends \WP_Upgrader
+{
+  public function __construct()
+  {
+    parent::__construct(new \Automatic_Upgrader_Skin());
+  }
+
+  public function upgrade(string $package_url): true|\WP_Error
+  {
+    \WP_Filesystem();
+
+    $this->init();
+
+    $result = $this->run([
+      'package'           => $package_url,
+      'destination'       => WPMU_PLUGIN_DIR . '/ionos-core',
+      'clear_destination' => true,
+      'clear_working'     => true,
+      'hook_extra'        => ['type' => 'plugin', 'action' => 'update'],
+    ]);
+
+    if (\is_wp_error($result)) {
+      return $result;
+    }
+
+    if (!$result) {
+      return new \WP_Error('mu_plugin_upgrade_failed', 'MU plugin upgrade failed with no result.');
+    }
+
+    return true;
+  }
+}
+
